new threaded session object.
